Add consistent navigation across all pages
- Create shared Layout component with header containing:
  - Tests link (to /#tests)
  - Results link (to /my-results)
  - Sign In / user menu dropdown
  - Theme toggle and GitHub icon

- Update UserMenu to show only Settings and Sign Out in dropdown
  (Results moved to main navigation)

- Remove "About" from navigation per user request

- Apply Layout to all pages and components:
  - Homepage, Settings, ResultsHistory, ResultDetail
  - PublicProfile, PublicResultDetail
  - All assessment and result components

- Remove duplicate ThemeProvider/ThemeToggle/GitHubIcon from App.tsx
  (now centralized in Layout.tsx)

- Update E2E tests to reflect navigation changes
