+++ frontend/App.tsx
@@ -1,5 +1,5 @@
 import { Routes, Route, Link, useParams } from 'react-router-dom';
-import { createContext, useContext, useEffect, useState } from 'react';
+import { useState } from 'react';
 import HexacoAssessment from './components/HexacoAssessment';
 import HexacoResults from './components/HexacoResults';
 import { HexacoResponse, calculateScores, DimensionScore } from './data/hexaco-scoring';
@@ -14,7 +14,7 @@ import { calculateEnneagramResult, type EnneagramResult } from './data/enneagram
 import { EnneagramResults } from './components/EnneagramResults';
 import { AuthProvider, useAuth } from './contexts/AuthContext';
 import { FeatureFlagsProvider, useFeatureFlags } from './contexts/FeatureFlagsContext';
-import { UserMenu } from './components/UserMenu';
+import { Layout, ThemeProvider } from './components/Layout';
 import { ResultsHistory } from './pages/ResultsHistory';
 import { ResultDetail } from './pages/ResultDetail';
 import { Settings } from './pages/Settings';
@@ -31,140 +31,11 @@ import MBTIResults from './components/MBTIResults';
 import RMETAssessment from './components/RMETAssessment';
 import RMETResults from './components/RMETResults';
 
-// Theme Context
-type Theme = 'dark' | 'light';
-
-const ThemeContext = createContext<{
-  theme: Theme;
-  toggleTheme: () => void;
-}>({
-  theme: 'dark',
-  toggleTheme: () => {},
-});
-
-function ThemeProvider({ children }: { children: React.ReactNode }) {
-  const [theme, setTheme] = useState<Theme>(() => {
-    if (typeof window !== 'undefined') {
-      const saved = localStorage.getItem('mesearch-theme');
-      return (saved as Theme) || 'dark';
-    }
-    return 'dark';
-  });
-
-  useEffect(() => {
-    const root = document.documentElement;
-    if (theme === 'light') {
-      root.classList.add('light');
-    } else {
-      root.classList.remove('light');
-    }
-    localStorage.setItem('mesearch-theme', theme);
-  }, [theme]);
-
-  const toggleTheme = () => {
-    setTheme((prev) => (prev === 'dark' ? 'light' : 'dark'));
-  };
-
-  return (
-    <ThemeContext.Provider value={{ theme, toggleTheme }}>
-      {children}
-    </ThemeContext.Provider>
-  );
-}
-
-function useTheme() {
-  return useContext(ThemeContext);
-}
-
-// Theme Toggle Button
-function ThemeToggle() {
-  const { theme, toggleTheme } = useTheme();
-
-  return (
-    <button
-      onClick={toggleTheme}
-      className="theme-toggle"
-      aria-label={`Switch to ${theme === 'dark' ? 'light' : 'dark'} mode`}
-    >
-      {theme === 'dark' ? (
-        // Sun icon for switching to light
-        <svg fill="none" viewBox="0 0 24 24" stroke="currentColor">
-          <path
-            strokeLinecap="round"
-            strokeLinejoin="round"
-            strokeWidth={1.5}
-            d="M12 3v1m0 16v1m9-9h-1M4 12H3m15.364 6.364l-.707-.707M6.343 6.343l-.707-.707m12.728 0l-.707.707M6.343 17.657l-.707.707M16 12a4 4 0 11-8 0 4 4 0 018 0z"
-          />
-        </svg>
-      ) : (
-        // Moon icon for switching to dark
-        <svg fill="none" viewBox="0 0 24 24" stroke="currentColor">
-          <path
-            strokeLinecap="round"
-            strokeLinejoin="round"
-            strokeWidth={1.5}
-            d="M20.354 15.354A9 9 0 018.646 3.646 9.003 9.003 0 0012 21a9.003 9.003 0 008.354-5.646z"
-          />
-        </svg>
-      )}
-    </button>
-  );
-}
-
-function GitHubIcon() {
-  return (
-    <a
-      href="https://github.com/emily-flambe/mesearch"
-      target="_blank"
-      rel="noopener noreferrer"
-      className="text-[var(--color-text-muted)] hover:text-[var(--color-champagne)] transition-colors duration-300"
-      aria-label="View on GitHub"
-    >
-      <svg
-        viewBox="0 0 24 24"
-        width="22"
-        height="22"
-        fill="currentColor"
-        aria-hidden="true"
-      >
-        <path d="M12 0c-6.626 0-12 5.373-12 12 0 5.302 3.438 9.8 8.207 11.387.599.111.793-.261.793-.577v-2.234c-3.338.726-4.033-1.416-4.033-1.416-.546-1.387-1.333-1.756-1.333-1.756-1.089-.745.083-.729.083-.729 1.205.084 1.839 1.237 1.839 1.237 1.07 1.834 2.807 1.304 3.492.997.107-.775.418-1.305.762-1.604-2.665-.305-5.467-1.334-5.467-5.931 0-1.311.469-2.381 1.236-3.221-.124-.303-.535-1.524.117-3.176 0 0 1.008-.322 3.301 1.23.957-.266 1.983-.399 3.003-.404 1.02.005 2.047.138 3.006.404 2.291-1.552 3.297-1.23 3.297-1.23.653 1.653.242 2.874.118 3.176.77.84 1.235 1.911 1.235 3.221 0 4.609-2.807 5.624-5.479 5.921.43.372.823 1.102.823 2.222v3.293c0 .319.192.694.801.576 4.765-1.589 8.199-6.086 8.199-11.386 0-6.627-5.373-12-12-12z" />
-      </svg>
-    </a>
-  );
-}
-
 function HomePage() {
   const { flags } = useFeatureFlags();
 
   return (
-    <div id="top" className="min-h-screen bg-[var(--color-bg-primary)] transition-colors duration-300">
-      {/* Header */}
-      <header className="border-b border-[var(--color-border-subtle)] bg-[var(--color-bg-primary)]/95 backdrop-blur-md sticky top-0 z-50 transition-colors duration-300">
-        <div className="mx-auto max-w-6xl px-6 py-5 flex items-center justify-between">
-          <a href="#top" className="font-display text-2xl font-semibold tracking-wide text-gold-gradient">
-            Mesearch
-          </a>
-          <nav className="flex items-center gap-6">
-            <a
-              href="#tests"
-              className="text-[var(--color-text-secondary)] hover:text-[var(--color-champagne)] transition-colors duration-300 text-sm tracking-wide uppercase"
-            >
-              Tests
-            </a>
-            <a
-              href="#about"
-              className="text-[var(--color-text-secondary)] hover:text-[var(--color-champagne)] transition-colors duration-300 text-sm tracking-wide uppercase"
-            >
-              About
-            </a>
-            <div className="w-px h-5 bg-[var(--color-border)]" />
-            <UserMenu />
-            <ThemeToggle />
-            <GitHubIcon />
-          </nav>
-        </div>
-      </header>
-
+    <Layout>
       <main>
         {/* Hero Section */}
         <section className="relative mx-auto max-w-6xl px-6 py-32 text-center overflow-hidden">
@@ -383,7 +254,7 @@ function HomePage() {
           <p className="text-[var(--color-text-muted)]/50 text-xs">&copy; 2026</p>
         </div>
       </footer>
-    </div>
+    </Layout>
   );
 }
 
@@ -460,21 +331,7 @@ function TestCard({
 // Generic placeholder for tests not yet implemented
 function ComingSoonTestPage() {
   return (
-    <div className="min-h-screen bg-[var(--color-bg-primary)] transition-colors duration-300">
-      {/* Header */}
-      <header className="border-b border-[var(--color-border-subtle)] bg-[var(--color-bg-primary)]/95 backdrop-blur-md transition-colors duration-300">
-        <div className="mx-auto max-w-6xl px-6 py-5 flex items-center justify-between">
-          <Link to="/" className="font-display text-2xl font-semibold tracking-wide text-gold-gradient">
-            Mesearch
-          </Link>
-          <div className="flex items-center gap-4">
-            <ThemeToggle />
-            <GitHubIcon />
-          </div>
-        </div>
-      </header>
-
-      {/* Content */}
+    <Layout>
       <main className="mx-auto max-w-2xl px-6 py-24 text-center">
         <div className="card-premium rounded-lg p-12">
           <div className="w-16 h-16 mx-auto mb-8 rounded-full border border-[var(--color-champagne)]/30 flex items-center justify-center">
@@ -506,7 +363,7 @@ function ComingSoonTestPage() {
           </Link>
         </div>
       </main>
-    </div>
+    </Layout>
   );
 }
 
@@ -580,20 +437,7 @@ function EnneagramTestPage() {
   };
 
   return (
-    <div className="min-h-screen bg-[var(--color-bg-primary)] transition-colors duration-300">
-      {/* Header */}
-      <header className="border-b border-[var(--color-border-subtle)] bg-[var(--color-bg-primary)]/95 backdrop-blur-md transition-colors duration-300">
-        <div className="mx-auto max-w-6xl px-6 py-5 flex items-center justify-between">
-          <Link to="/" className="font-display text-2xl font-semibold tracking-wide text-gold-gradient">
-            Mesearch
-          </Link>
-          <div className="flex items-center gap-4">
-            <ThemeToggle />
-            <GitHubIcon />
-          </div>
-        </div>
-      </header>
-
+    <Layout>
       <main className="mx-auto max-w-3xl px-6 py-12">
         {phase === 'intro' && (
           <div className="text-center">
@@ -721,7 +565,7 @@ function EnneagramTestPage() {
           <EnneagramResults result={result} onRetake={handleRetake} />
         )}
       </main>
-    </div>
+    </Layout>
   );
 }
