+++ frontend/components/ECRAssessment.tsx
@@ -1,5 +1,6 @@
 import { useState, useEffect, useCallback } from 'react';
 import { useNavigate, Link } from 'react-router-dom';
+import { Layout } from './Layout';
 import { items, likertScale, type Item, type LikertValue } from '../data/ecr-items';
 import {
   type Response,
@@ -134,8 +135,7 @@ export default function ECRAssessment() {
   // Render intro phase
   if (phase === 'intro') {
     return (
-      <div className="min-h-screen bg-[var(--color-bg-primary)] transition-colors duration-300">
-        <Header />
+      <Layout>
         <main className="mx-auto max-w-2xl px-6 py-16">
           <div className="card-premium rounded-lg p-10">
             <div className="text-center mb-8">
@@ -213,15 +213,14 @@ export default function ECRAssessment() {
             </div>
           </div>
         </main>
-      </div>
+      </Layout>
     );
   }
 
   // Render completion phase
   if (phase === 'complete') {
     return (
-      <div className="min-h-screen bg-[var(--color-bg-primary)] transition-colors duration-300">
-        <Header />
+      <Layout>
         <main className="mx-auto max-w-2xl px-6 py-16">
           <div className="card-premium rounded-lg p-10 text-center">
             <div className="w-16 h-16 mx-auto mb-6 rounded-full border-2 border-[var(--color-champagne)] flex items-center justify-center">
@@ -257,7 +256,7 @@ export default function ECRAssessment() {
             </button>
           </div>
         </main>
-      </div>
+      </Layout>
     );
   }
 
@@ -356,20 +355,5 @@ export default function ECRAssessment() {
         </div>
       </footer>
     </div>
-  );
-}
-
-function Header() {
-  return (
-    <header className="border-b border-[var(--color-border-subtle)] bg-[var(--color-bg-primary)]/95 backdrop-blur-md transition-colors duration-300">
-      <div className="mx-auto max-w-6xl px-6 py-5 flex items-center justify-between">
-        <Link
-          to="/"
-          className="font-display text-2xl font-semibold tracking-wide text-gold-gradient"
-        >
-          Mesearch
-        </Link>
-      </div>
-    </header>
   );
 }
