+++ frontend/components/MFQAssessment.tsx
@@ -1,5 +1,6 @@
 import { useState, useEffect, useCallback } from 'react';
 import { useNavigate, Link } from 'react-router-dom';
+import { Layout } from './Layout';
 import { items, shuffleItems, type Item, citation } from '../data/mfq-items';
 import {
   type Response,
@@ -162,8 +163,7 @@ export default function MFQAssessment() {
   // Render intro phase
   if (phase === 'intro') {
     return (
-      <div className="min-h-screen bg-[var(--color-bg-primary)] transition-colors duration-300">
-        <Header />
+      <Layout>
         <main className="mx-auto max-w-2xl px-6 py-16">
           <div className="card-premium rounded-lg p-10">
             <div className="text-center mb-8">
@@ -235,15 +235,14 @@ export default function MFQAssessment() {
             </div>
           </div>
         </main>
-      </div>
+      </Layout>
     );
   }
 
   // Render completion phase
   if (phase === 'complete') {
     return (
-      <div className="min-h-screen bg-[var(--color-bg-primary)] transition-colors duration-300">
-        <Header />
+      <Layout>
         <main className="mx-auto max-w-2xl px-6 py-16">
           <div className="card-premium rounded-lg p-10 text-center">
             <div className="w-16 h-16 mx-auto mb-6 rounded-full border-2 border-[var(--color-champagne)] flex items-center justify-center">
@@ -283,7 +282,7 @@ export default function MFQAssessment() {
             </button>
           </div>
         </main>
-      </div>
+      </Layout>
     );
   }
 
@@ -394,20 +393,5 @@ export default function MFQAssessment() {
         </div>
       </footer>
     </div>
-  );
-}
-
-function Header() {
-  return (
-    <header className="border-b border-[var(--color-border-subtle)] bg-[var(--color-bg-primary)]/95 backdrop-blur-md transition-colors duration-300">
-      <div className="mx-auto max-w-6xl px-6 py-5 flex items-center justify-between">
-        <Link
-          to="/"
-          className="font-display text-2xl font-semibold tracking-wide text-gold-gradient"
-        >
-          Mesearch
-        </Link>
-      </div>
-    </header>
   );
 }
