+++ frontend/pages/ResultDetail.tsx
@@ -1,7 +1,7 @@
 import { useState, useEffect } from 'react';
 import { Link, useParams, useNavigate } from 'react-router-dom';
 import { useAuth } from '../contexts/AuthContext';
-import { UserMenu } from '../components/UserMenu';
+import { Layout } from '../components/Layout';
 import BigFiveResults from '../components/BigFiveResults';
 import { EnneagramResults } from '../components/EnneagramResults';
 import HexacoResults from '../components/HexacoResults';
@@ -116,8 +116,7 @@ export function ResultDetail() {
 
   if (!user) {
     return (
-      <div className="min-h-screen bg-[var(--color-bg-primary)] transition-colors duration-300">
-        <Header />
+      <Layout>
         <main className="mx-auto max-w-4xl px-6 py-12">
           <div className="card-premium rounded-lg p-12 text-center">
             <div className="w-16 h-16 mx-auto mb-6 rounded-full border border-[var(--color-champagne)]/30 flex items-center justify-center">
@@ -131,14 +130,13 @@ export function ResultDetail() {
             </p>
           </div>
         </main>
-      </div>
+      </Layout>
     );
   }
 
   if (error) {
     return (
-      <div className="min-h-screen bg-[var(--color-bg-primary)] transition-colors duration-300">
-        <Header />
+      <Layout>
         <main className="mx-auto max-w-4xl px-6 py-12">
           <div className="card-premium rounded-lg p-12 text-center">
             <div className="w-16 h-16 mx-auto mb-6 rounded-full border border-red-500/30 flex items-center justify-center">
@@ -156,14 +154,13 @@ export function ResultDetail() {
             </Link>
           </div>
         </main>
-      </div>
+      </Layout>
     );
   }
 
   if (!result) {
     return (
-      <div className="min-h-screen bg-[var(--color-bg-primary)] transition-colors duration-300">
-        <Header />
+      <Layout>
         <main className="mx-auto max-w-4xl px-6 py-12">
           <div className="card-premium rounded-lg p-12 text-center">
             <h2 className="font-display text-xl text-[var(--color-text-primary)] mb-2">Result Not Found</h2>
@@ -178,15 +175,14 @@ export function ResultDetail() {
             </Link>
           </div>
         </main>
-      </div>
+      </Layout>
     );
   }
 
   const retakeUrl = testUrls[result.test_type] || `/test/${result.test_type}`;
 
   return (
-    <div className="min-h-screen bg-[var(--color-bg-primary)] transition-colors duration-300">
-      <Header />
+    <Layout>
       <main className="mx-auto max-w-4xl px-6 py-12">
         {/* Back link */}
         <Link
@@ -300,28 +296,7 @@ export function ResultDetail() {
           </Link>
         </div>
       </main>
-    </div>
-  );
-}
-
-function Header() {
-  return (
-    <header className="border-b border-[var(--color-border-subtle)] bg-[var(--color-bg-primary)]/95 backdrop-blur-md transition-colors duration-300">
-      <div className="mx-auto max-w-6xl px-6 py-5 flex items-center justify-between">
-        <Link to="/" className="font-display text-2xl font-semibold tracking-wide text-gold-gradient">
-          Mesearch
-        </Link>
-        <nav className="flex items-center gap-6">
-          <Link
-            to="/my-results"
-            className="text-[var(--color-text-secondary)] hover:text-[var(--color-champagne)] transition-colors text-sm"
-          >
-            My Results
-          </Link>
-          <UserMenu />
-        </nav>
-      </div>
-    </header>
+    </Layout>
   );
 }
